+++ lib/ops/gaussian_blur/index.js
@@ -6,9 +6,8 @@
  * All rights reserved.
  */
 
-import RegisterOperation from '../../program/operation_register';
-import kernel from './kernel.glsl';
-import Tensor from '../../program/tensor';
+import Convolutiion from '../conv2d';
+import { gaussianBlur } from '../conv2d/kernels';
 import * as utils from '../../utils';
 
 /**
@@ -34,39 +33,5 @@ export default (tSrc, kernelSize = 3, sigma = 3) => {
     'Sigma should be greater then 0',
   );
 
-  return new RegisterOperation('GaussianBlur')
-    .Input('tSrc', tSrc.dtype)
-    .Input('tKernel', 'float32')
-    .Output(tSrc.dtype)
-    .LoadChunk('pickValue')
-    .Uniform('uWidth', 'float', tSrc.shape[0])
-    .Uniform('uHeight', 'float', tSrc.shape[1])
-    .Constant('KERNEL_SIZE', kernelSize)
-    .PreCompile((op) => {
-      const dstKernel = new Tensor('float32', [kernelSize, kernelSize]);
-      const mean = kernelSize / 2;
-      const resultKernel = new Tensor('float32', [kernelSize, kernelSize, 4]);
-      let sum = 0.0;
-
-      for (let x = 0; x < kernelSize; x += 1) {
-        for (let y = 0; y < kernelSize; y += 1) {
-          const v = Math.exp(-0.5 * (((x - mean) / sigma) ** 2
-            + ((y - mean) / sigma) ** 2)) / (2 * Math.PI * sigma * sigma);
-
-          dstKernel.set(x, y, v);
-          sum += dstKernel.get(x, y);
-        }
-      }
-
-      // Normalize the kernel
-      for (let x = 0; x < kernelSize; x += 1) {
-        for (let y = 0; y < kernelSize; y += 1) {
-          resultKernel.set(x, y, 3, dstKernel.get(x, y) / sum);
-        }
-      }
-
-      op.assignInput('tKernel', resultKernel);
-    })
-    .GLSLKernel(kernel)
-    .Compile({ tSrc });
+  return Convolutiion(tSrc, gaussianBlur(sigma));
 };
